Migrate to using CFFI 1.0
* Move everything under src/ to ensure we test against the installed
  library.
* Create a build_bcrypt.py script which will build _bcrypt.so.
* Refactor to utilize the new _bcrypt.so instead of implicit compile.
